refactor: extract comment-reply logic into shared handler to eliminate self-referencing HTTP fetch
